Add tuition widget
