+++ Campus-iOS/TuitionComponent/View/TuitionView.swift
@@ -15,7 +15,7 @@ struct TuitionView: View {
         List {
             VStack(alignment: .center) {
                 Spacer(minLength: 0.10 * UIScreen.main.bounds.width)
-                TuitionCard(tuition: self.viewModel.tuition ?? Tuition(deadline: Date(), semester: "Unknown", semesterID: "0", amount: 0))
+                TuitionCard(tuition: self.viewModel.tuition ?? Tuition.unknown)
             }
             .listRowBackground(Color(.systemGroupedBackground))
         }
